LayoutTests/userscripts/window-onerror-for-isolated-world-2-expected.txt

CONSOLE MESSAGE: line 30: Error: Error in main world inline script.
Test that window.onerror and "error" event listeners from isolated world are invoked for uncaught exceptions in user scripts running in isolate worlds as well as for exceptions in the main world.Bug 8519.

user script window.onerror: Error: Error in user script inline script. at undefined:31
user script error event listener: Error: Error in user script inline script. at undefined:31
user script window.onerror: Error: Error in main world load handler. at window-onerror-for-isolated-world-2.html:27
user script error event listener: Error: Error in main world load handler. at window-onerror-for-isolated-world-2.html:27
user script window.onerror: Error: Error in user script load handler. at undefined:28
user script error event listener: Error: Error in user script load handler. at undefined:28
user script window.onerror: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-2.html:25
user script error event listener: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-2.html:25
user script window.onerror: Error: Error in user script setTimeout callback. at undefined:26
user script error event listener: Error: Error in user script setTimeout callback. at undefined:26
